+++ apps/web/CHANGELOG.md
@@ -1,5 +1,18 @@
 # web
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @faiz-ui/react@1.1.2
+
+## 1.0.5
+
+### Patch Changes
+
+- @faiz-ui/react@1.1.1
+
 ## 1.0.4
 
 ### Patch Changes