+++ packages/components/landing-card/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @faiz-ui/landing-card
 
+## 1.0.6
+
+### Patch Changes
+
+- updating changelog
+- Updated dependencies
+  - @faiz-ui/theme@1.1.2
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @faiz-ui/theme@1.1.1
+
 ## 1.0.1
 
 ### Patch Changes